+++ engine/static/input/src/keyboard.cpp
@@ -270,7 +270,7 @@ namespace argus {
         }
 
         KeyboardScancode scancode = _translate_glfw_keycode(glfw_keycode);
-        KeyboardModifiers mod = _translate_glfw_keymod(glfw_mods);
+        KeyboardModifiers mod = _translate_glfw_keymod(static_cast<uint16_t>(glfw_mods));
 
         dispatch_event<KeyboardEvent>(key_event_type, scancode, mod);
     }
